requirements.txt
# AI packages
langchain_openai
langchain_community
openai
python-dotenv
langchain_groq
pypdf
pymupdf
streamlit
openpyxl